# api/openapi.yaml
openapi: 3.1.0
info:
  title: FasalSetu API
  version: 1.0.0
  summary: Revenue-first, multilingual agro-advisory API (digital public good).
  description: |
    Every recommendation returned by this API has passed through the revenue layer
    (money impact or an explicit do-nothing) and the compliance guardrail, and has
    been written to the append-only event log. Money figures are ESTIMATES and always
    carry their assumptions.
  license: { name: Apache-2.0, identifier: Apache-2.0 }
servers:
  - url: https://api.example.org/v1
...
security:
  - bearerAuth: []
paths:
...
  /market/prices:
    get:
      tags: [specialists]
      parameters:
        - { name: crop, in: query, required: true, schema: { type: string } }
        - { name: market, in: query, schema: { type: string } }
      summary: Prices from the country pack's source vs support price, with sell/hold view.
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MarketView' }
  /schemes/search:
    get:
      tags: [specialists]
      parameters:
        - { name: q, in: query, required: true, schema: { type: string } }
        - { name: language, in: query, schema: { type: string } }
      summary: Semantic search over the country's schemes.
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Scheme' }
...
  /webhooks/whatsapp:
    post:
      tags: [delivery]
      summary: Inbound WhatsApp message/status. Replies like "1 = done" become response events.
      security: [{ webhookSignature: [] }]
      requestBody: { content: { application/json: { schema: { type: object } } } }
      responses: { '200': { description: OK } }
  /webhooks/sms:
    post:
      tags: [delivery]
      summary: Inbound SMS reply or delivery receipt.
      security: [{ webhookSignature: [] }]
      requestBody: { content: { application/json: { schema: { type: object } } } }
      responses: { '200': { description: OK } }
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer }
    webhookSignature: { type: apiKey, in: header, name: X-Signature }
...
  schemas:
...
    MarketView:
      type: object
      properties:
        crop: { type: string }
        price_per_quintal: { type: number }
        support_price: { type: [number, 'null'] }
        source: { type: string }
        observed_at: { type: string, format: date-time }
        advice: { type: string, enum: [sell, hold, unknown] }
    Scheme:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        summary: { type: string }
        eligibility: { type: string }
        url: { type: string }
        score: { type: number }
        benefit_type:
          type: string
          description: "Optional. e.g. SUBSIDY, DIRECT_TRANSFER, INSURANCE, CREDIT, MSP"
        benefit_amount:
          type: [string, 'null']
          description: "Optional. Human-readable benefit amount or range, e.g. '₹6,000/year' or 'Up to 50%'."
        level:
          type: string
          description: "Optional. CENTRAL or STATE."
          enum: [CENTRAL, STATE]
        state:
          type: [string, 'null']
          description: "Optional. State name if level=STATE; null for central schemes."
...